syncing goodVertices and primaryVertexFilter

+++ skim/python/skim_4T.py
@@ -211,7 +211,7 @@ process.eeBadScFilter.taggingMode = cms.bool(True)
 process.primaryVertexFilter = cms.EDFilter("GoodVertexFilter",
                                            vertexCollection = cms.InputTag('offlinePrimaryVertices'),
                                            minimumNDOF = cms.uint32(4) ,
-                                           maxAbsZ = cms.double(15),
+                                           maxAbsZ = cms.double(24),
                                            maxd0 = cms.double(2)
                                            )
